requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.20
langsmith>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-docx>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
matplotlib>=3.8.0
pandas>=2.1.0
numpy>=1.26.0
seaborn>=0.13.0
Pillow>=10.2.0
requests>=2.31.0
tavily-python>=0.3.0

